requirements.txt
sanic~=20.12.2
SQLAlchemy~=1.3.23
aiogram~=2.12
secure~=0.2.1
APScheduler~=3.7.0
databases~=0.4.2
python-dotenv~=0.15.0
